glade-gtk: Maintain icon-name attribute for GtkIconFactory.
The attribute was dropped during loading and not written to the output XML
on saving. With this change it's preserved up through output XML generation.

Alongside, as icon-name can be used instead of the filename attribute the
filename attribute changed from required to optional.

NOTE: This does not change the UI yet which does still not handle icon-name, so
for now icon-name must be changed in the XML directly.